requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via starlette
click==8.1.7
    # via uvicorn
exceptiongroup==1.2.2
    # via
    #   anyio
    #   pytest
faiss-gpu==1.7.2
    # via test-time-classifier (pyproject.toml)
fastapi==0.115.5
    # via test-time-classifier (pyproject.toml)
filelock==3.16.1
    # via
    #   torch
    #   triton
fsspec==2024.10.0
    # via torch
h11==0.14.0
    # via uvicorn
idna==3.10
    # via anyio
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via torch
joblib==1.4.2
    # via scikit-learn
loguru==0.7.2
    # via test-time-classifier (pyproject.toml)
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==1.26.4
    # via
    #   test-time-classifier (pyproject.toml)
    #   pandas
    #   scikit-learn
    #   scipy
    #   torchvision
nvidia-cublas-cu12==12.4.5.8
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127
    # via torch
nvidia-cuda-runtime-cu12==12.4.127
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.2.1.3
    # via torch
nvidia-curand-cu12==10.3.5.147
    # via torch
nvidia-cusolver-cu12==11.6.1.9
    # via torch
nvidia-cusparse-cu12==12.3.1.170
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.21.5
    # via torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127
    # via torch
packaging==24.2
    # via pytest
pandas==2.2.3
    # via test-time-classifier (pyproject.toml)
pillow==11.0.0
    # via torchvision
pluggy==1.5.0
    # via pytest
pydantic==2.9.2
    # via fastapi
pydantic-core==2.23.4
    # via pydantic
pytest==8.3.3
    # via test-time-classifier (pyproject.toml)
python-dateutil==2.9.0.post0
    # via pandas
python-multipart==0.0.17
    # via test-time-classifier (pyproject.toml)
pytz==2024.2
    # via pandas
scikit-learn==1.5.2
    # via test-time-classifier (pyproject.toml)
scipy==1.14.1
    # via scikit-learn
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.41.2
    # via fastapi
sympy==1.13.1
    # via torch
threadpoolctl==3.5.0
    # via scikit-learn
tomli==2.1.0
    # via pytest
torch==2.5.1
    # via
    #   test-time-classifier (pyproject.toml)
    #   torchvision
torchvision==0.20.1
    # via test-time-classifier (pyproject.toml)
tqdm==4.67.0
    # via test-time-classifier (pyproject.toml)
triton==3.1.0
    # via torch
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   torch
    #   uvicorn
tzdata==2024.2
    # via pandas
uvicorn==0.32.0
    # via test-time-classifier (pyproject.toml)
